+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Asynchronous messaging for .NET Changelog
 
+## <a name="3.2.0"></a> 3.2.0 (2021-03-23)
+
+Improved message queues
+
+### Features
+* **queues** Added CallbackMessageReceiver to wrap callbacks into IMessageReceiver interface
+* **queues** Addded IMessageConnection interface
+* **build** Set config params and references to created queues in MessageQueueFactory
+* **queues** Added CheckOpen method to MessageQueue
+* **queues** Added JSON serialization for MessageEnvelop
+* **build** Added IMessageQueueFactory interface
+* **build** Added MessageQueueFactory abstract class
+
 ## <a name="3.1.1"></a> 3.1.1 (2020-06-26)
 
 ### Features